add jenkins jobs model, refactoring view, models struct,
