+++ angular-admin-ui/src/app/models/jenkins-job.ts
@@ -1,0 +1,7 @@
+export class JenkinsJob {
+  id: number;
+  name: string;
+  jenkins_group_id: number;
+  jenkins_job_perent_id: number;
+  gitlab_project_id: number;
+}
