+++ server/core/models/__init__.py
@@ -1,1 +1,4 @@
 #Database Layer
+class RecordNotFound(Exception):
+    """Requested record in database was not found"""
+    pass
