+++ server/core/models/jenkins_jobs.py
@@ -1,19 +1,18 @@
-import enum
-import hashlib
-import json
 from sqlalchemy import desc
 
 from server.core.common import LoggingMixin
 from server.core.models import RecordNotFound
 
 
-class JenkinsGroup(object):
-    """Data class for JenkinsGroup"""
+class JenkinsJob(object):
+    """Data class for JenkinsJob"""
 
     def __init__(self):
         self.id = None
         self.name = None
-        self.jobs_base_path = None
+        self.jenkins_group_id = None
+        self.jenkins_job_perent_id = None
+        self.gitlab_project_id = None
 
     @property
     def values(self):
@@ -21,62 +20,60 @@ class JenkinsGroup(object):
         return result
 
     @staticmethod
-    def make(name, jobs_base_path):
-        obj = JenkinsGroup()
+    def make(name, jenkins_group_id, jenkins_job_perent_id, gitlab_project_id):
+        obj = JenkinsJob()
         obj.name = name
-        obj.jobs_base_path = jobs_base_path
+        obj.jobs_base_path = jenkins_group_id
+        obj.jenkins_job_perent_id = jenkins_job_perent_id
+        obj.gitlab_project_id = gitlab_project_id
+
         return obj
 
 
-class JenkinsGroupManager(LoggingMixin):
-    """ Class for managment data in jenkins_groups table"""
+class JenkinsJobManager(LoggingMixin):
+    """ Class for managment data in jenkins_jobs table"""
 
     def __init__(self, marker, db_pool, tables):
         self._marker = marker
         self.db_pool = db_pool
-        self._jenkins_groups = tables['jenkins_groups']
+        self._jenkins_jobs = tables['jenkins_jobs']
 
-    async def search(self, name, limit=150):
+    async def find_by_groud_id(self, group_id):
         """
-        Search jenkins_groups by fields
+        Search jenkins_groups by group_id
 
         Args:
-            name - group name
+            group_id - group id
         Return:
             list[JenkinsGroup obj, JenkinsGroup obj]
 
         Exceptions:
                 pymysql.err
         """
-        self._logging_debug("Search: %s" % (name))
+        self._logging_debug("Find by group id: %d" % group_id)
 
-        colums = self._jenkins_groups.c
+        columns = self._jenkins_jobs.c
         async with self.db_pool.acquire() as conn:
-            q = self._jenkins_groups.select()
-
-            if name:
-                q = q.where(colums.name.startswith(name))
-
-            if limit:
-                q = q.limit(limit)
-
-            q = q.order_by(desc(colums.name))
-
+            q = self._jenkins_jobs.select().\
+                where(columns.jenkins_group_id == group_id).\
+                order_by(columns.id)
+            
             self._logging_debug(q)
+
             result = await conn.execute(q)
             rows = await result.fetchall()
             list_data_obj = await self._mapping_from_tuple(rows)
             return list_data_obj
 
-    async def get(self, grop_id):
+    async def get(self, job_id):
         """
-        Get one JenkinsGroup object by id field
+        Get one JenkinsJob object by id field
 
         Args:
-            grop_id - value id field
+            job_id - value id field
 
         Return:
-            JenkinsGroup object
+            JenkinsJob object
 
         Exceptions:
                 RecordNotFound
@@ -84,37 +81,37 @@ class JenkinsGroupManager(LoggingMixin):
                 and anower from sqlalchemy
         """
         async with self.db_pool.acquire() as conn:
-            q = self._jenkins_groups.select().where(self._jenkins_groups.c.id == grop_id)
+            q = self._jenkins_jobs.select().where(self._jenkins_jobs.c.id == job_id)
             self._logging_debug(q)
             result = await conn.execute(q)
             row = await result.fetchone()
             if row is None:
                 msg = "JenkinsGroup with id: {} does not exists"
-                raise RecordNotFound(msg.format(grop_id))
+                raise RecordNotFound(msg.format(job_id))
             obj = await self._mapping_row_to_obj(row)
             self._logging_debug(obj)
             return obj
 
-    async def create(self, jenkins_group):
+    async def create(self, jenkins_job):
         """
         Create JenkinsGroup object
 
         Args:
-            jenkins_group - JenkinsGroup object
+            jenkins_job - JenkinsJob object
 
         Return:
-            JenkinsGroup object
+            JenkinsJob object
 
         Exceptions:
                 pymysql.err
                 and anower from sqlalchemy
         """
         async with self.db_pool.acquire() as conn:
-            self._logging_debug(jenkins_group.values)
+            self._logging_debug(jenkins_job.values)
             trans = await conn.begin()
             try:
-                q = self._jenkins_groups.insert(). \
-                    values(jenkins_group.values)
+                q = self._jenkins_jobs.insert(). \
+                    values(jenkins_job.values)
                 result = await conn.execute(q)
                 await trans.commit()
                 self._logging_debug('Commit')
@@ -124,12 +121,12 @@ class JenkinsGroupManager(LoggingMixin):
                 await trans.rollback()
                 raise
 
-    async def update(self, jenkins_group):
+    async def update(self, jenkins_job):
         """
         Update values
 
         Args:
-            jenkins_group - JenkinsGroup
+            jenkins_group - JenkinsJob
 
         Return:
             None
@@ -141,9 +138,9 @@ class JenkinsGroupManager(LoggingMixin):
         async with self.db_pool.acquire() as conn:
             trans = await conn.begin()
             try:
-                q = self._jenkins_groups. \
-                    update(self._jenkins_groups.c.id == jenkins_group.id). \
-                    values(jenkins_group.values)
+                q = self._jenkins_jobs. \
+                    update(self._jenkins_jobs.c.id == jenkins_job.id). \
+                    values(jenkins_job.values)
                 self._logging_debug(q)
                 await conn.execute(q)
                 await trans.commit()
@@ -152,12 +149,12 @@ class JenkinsGroupManager(LoggingMixin):
                 await trans.rollback()
                 raise
 
-    async def delete(self, grop_id):
+    async def delete(self, job_id):
         """
         Delete row
 
         Args:
-            grop_id - JenkinsGroup id
+            job_id - JenkinsJob id
 
         Return:
             None
@@ -169,7 +166,7 @@ class JenkinsGroupManager(LoggingMixin):
         async with self.db_pool.acquire() as conn:
             trans = await conn.begin()
             try:
-                q = self._jenkins_groups.delete(self._jenkins_groups.c.id == grop_id)
+                q = self._jenkins_jobs.delete(self._jenkins_jobs.c.id == job_id)
                 self._logging_debug(q)
                 await conn.execute(q)
                 await trans.commit()
@@ -180,16 +177,16 @@ class JenkinsGroupManager(LoggingMixin):
 
     # private methods
     def _columns(self):
-        return self._jenkins_groups.c.keys()
+        return self._jenkins_jobs.c.keys()
 
     async def _mapping_row_to_obj(self, row):
         """
-        Mapping row to JenkinsGroup object
+        Mapping row to JenkinsJob object
 
         Return:
             JenkinsGroup object
         """
-        data_obj = JenkinsGroup()
+        data_obj = JenkinsJob()
         for column in self._columns():
             data_obj.__setattr__(column, row[column])
 
@@ -197,10 +194,10 @@ class JenkinsGroupManager(LoggingMixin):
 
     async def _mapping_from_tuple(self, rows):
         """
-        Mapping tuple(row,row) to list [JenkinsGroup obj, JenkinsGroup obj]
+        Mapping tuple(row,row) to list [JenkinsJob obj, JenkinsJob obj]
 
         Return:
-            list[JenkinsGroup obj]
+            list[JenkinsJob obj]
         """
         result = []
         for row in rows:
