+++ server/core/services/gitlab_merge_service.py
@@ -1,7 +1,7 @@
 from aiohttp.client_exceptions import ClientResponseError, ClientConnectorError
 
 from server.core.common import LoggingMixin
-from server.core.models.delayed_tasks import DelayedTaskManager, DelayedTaskStatus, DelayedTaskType, DelayedTask, RecordNotFound
+from server.core.models.delayed_tasks import DelayedTaskManager, DelayedTaskStatus, DelayedTask, RecordNotFound
 from server.core.clients.jenkins_client import JenkinsClient
 from server.core.clients.gitlab_client import GitLabClient, GitLabMergeState, GitLabMerge
